+++ main.go
@@ -17,7 +17,7 @@ var assets embed.FS
 var appIcon []byte
 
 // Keep in sync with package.json and build/config.yml
-const appVersion = "2.1.0"
+const appVersion = "3.0.0"
 
 func main() {
 	backend.FixPath()